+++ packages/client/src/StreamrClient.ts
@@ -199,6 +199,7 @@ export class StreamrClient {
                 streamPartId,
                 options.resend,
                 this.resends,
+                this.streamStorageRegistry,
                 this.loggerFactory,
                 this.config
             )